feat

+++ scheduler/scheduler.h
@@ -36,19 +36,19 @@ typedef enum {
 typedef struct {
   u32        id;           // 任务ID
   u32        priority;     // 任务优先级, 数值越小优先级越高
-  u64        exec_freq;    // 执行频率
-  u64        exec_cnt_max; // 执行次数
-  u64        delay;        // 初始延时
+  u32        exec_freq;    // 执行频率
+  u32        exec_cnt_max; // 执行次数
+  u32        delay;        // 初始延时
   sched_cb_f f_cb;         // 回调函数
   void      *arg;          // 回调参数
 } sched_task_cfg_t;
 
 typedef struct {
   sched_task_state_e state;
-  u64                exec_cnt;
-  u64                elapsed;
-  u64                create_ts;
-  u64                next_exec_ts;
+  u32                exec_cnt;
+  u32                elapsed;
+  u32                create_ts;
+  u32                next_exec_ts;
 } sched_task_status_t;
 
 typedef struct {
@@ -62,7 +62,7 @@ typedef struct {
   f32          elapsed_us;
   sched_task_t tasks[SCHED_TASK_MAX];
   u32          tasks_num;
-  u64          curr_ts;
+  u32          curr_ts;
   rb_root_t    ready_tree;
 } sched_lo_t;
